chore(db): create wallet_live database and verify PostgreSQL setup

+++ src/models.rs
@@ -1,8 +1,8 @@
 use serde::{ Serialize, Deserialize};
 
-#[derive(Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Asset {
-  pub id: i64,
+  pub id: i32,
   pub name: String,
   pub unit_value: f64,
 }
